+++ src/main/java/org/purescript/psi/exports/ExportedValueReference.kt
@@ -1,10 +1,9 @@
-package org.purescript.psi.`var`
+package org.purescript.psi.exports
 
 import com.intellij.psi.PsiElementResolveResult.createResults
 import com.intellij.psi.PsiNamedElement
 import com.intellij.psi.PsiReferenceBase
 import com.intellij.psi.ResolveResult
-import org.purescript.psi.PSExportedValue
 
 class ExportedValueReference(exportedValue: PSExportedValue) : PsiReferenceBase.Poly<PSExportedValue>(
     exportedValue,
