+++ src/main/java/org/purescript/psi/exports/PSExportedDataMember.kt
@@ -1,5 +1,6 @@
-package org.purescript.psi
+package org.purescript.psi.exports
 
 import com.intellij.lang.ASTNode
+import org.purescript.psi.PSPsiElement
 
 class PSExportedDataMember(node: ASTNode) : PSPsiElement(node)
